seperated out the session manager class into its own file
